feat: Add tests for getTotalFeeUsd and mapWalletToAssetAndFeeList methods in FeeService; remove old test file
