+++ src/cashier_frontend_new/src/modules/transactionCart/state/txCartStore.svelte.ts
@@ -342,5 +342,4 @@ export class TransactionCartStore<T extends TransactionSource> {
       return Err((e as Error).message);
     }
   }
-
 }
